+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Tests for `TransformUtils`
 - Tests for `ShapeCreator` (Box, BoxLine, Cylinder, Pointer, Sphere, Arrow, Gizmo)
 - Tests for `DirectShapeUtils`
+- Tests for `TessellatedShapeCreator`
 
 ## [0.1.0] / 2023-10-01
 ### Quaternion